Now var size is available in asm

+++ soft/expr/parser.h
@@ -13,4 +13,4 @@ Node* declare();
 Node* stmt();
 bool stmts(std::vector<Node*>& statements);
 
-bool isDeclared(std::string var_name, int* var_size = 0);
+bool isDeclared(std::string var_name, uint8_t* var_size = 0);
