+++ soft/expr/var.cc
@@ -1,6 +1,6 @@
 #include "var.h"
 #include "var_size.h"
 
-int Var::size() const {
+uint8_t Var::size() const {
   return ::var_size(data_type, false);
 }
